Credit cards

+++ My Assets/MyAssetsApp.swift
@@ -23,10 +23,11 @@ struct MyAssetsApp: App {
 //            DeleterView()
             RootTabView()
         }
-        .modelContainer(for: [Asset.self, Debt.self, Stock.self, UpcomingSpend.self, Income.self, Expense.self])
-//        #if DEBUG
-//        .modelContainer(previewContainer)
-//        #endif
+        .defaultSize(width: 500, height: 800)
+//        .modelContainer(for: [Asset.self, Debt.self, Stock.self, UpcomingSpend.self, Income.self, Expense.self, CreditCard.self])
+        #if DEBUG
+        .modelContainer(previewContainer)
+        #endif
         
         #if os(macOS)
         Settings {
@@ -45,11 +46,28 @@ struct DeleterView: View {
     @Query var nonAssetIncome: [Income]
     @Query var nonDebtExpenses: [Expense]
     var body: some View {
-        EmptyView()
+        Text("Deleting preview data...")
             .onAppear {
-                let items: [any PersistentModel] = nonStockAssets + debts + nonAssetIncome + nonDebtExpenses
-                for item in items {
-                    modelContext.delete(item)
+                for item in nonStockAssets {
+                    if previewAssets.contains(where: { $0.name == item.name && $0.currentValue == item.currentValue }) {
+                        modelContext.delete(item)
+                        print("Delete \(item.name ?? "")")
+                    }
+                }
+                for item in debts {
+                    if previewDebts.contains(where: { $0.name == item.name && $0.currentValue == item.currentValue }) {
+                        modelContext.delete(item)
+                    }
+                }
+                for item in nonAssetIncome {
+                    if previewIncome.contains(where: { $0.name == item.name && $0.monthlyEarnings == item.monthlyEarnings }) {
+                        modelContext.delete(item)
+                    }
+                }
+                for item in nonDebtExpenses {
+                    if previewExpenses.contains(where: { $0.name == item.name && $0.monthlyCost == item.monthlyCost }) {
+                        modelContext.delete(item)
+                    }
                 }
             }
     }
